refactor: move modal to own file
